Refactor common prometheus stat code
