+++ controller/api/public/stat_summary.go
@@ -3,8 +3,6 @@ package public
 import (
 	"context"
 	"fmt"
-	"math"
-	"time"
 
 	proto "github.com/golang/protobuf/proto"
 	"github.com/linkerd/linkerd2/controller/api/util"
@@ -17,13 +15,6 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 )
-
-type promType string
-type promResult struct {
-	prom promType
-	vec  model.Vector
-	err  error
-}
 
 type resourceResult struct {
 	res *pb.StatTable
@@ -44,17 +35,7 @@ type rKey struct {
 const (
 	reqQuery             = "sum(increase(response_total%s[%s])) by (%s, classification, tls)"
 	latencyQuantileQuery = "histogram_quantile(%s, sum(irate(response_latency_ms_bucket%s[%s])) by (le, %s))"
-
-	promRequests   = promType("QUERY_REQUESTS")
-	promLatencyP50 = promType("0.5")
-	promLatencyP95 = promType("0.95")
-	promLatencyP99 = promType("0.99")
-
-	namespaceLabel    = model.LabelName("namespace")
-	dstNamespaceLabel = model.LabelName("dst_namespace")
 )
-
-var promTypes = []promType{promRequests, promLatencyP50, promLatencyP95, promLatencyP99}
 
 type podStats struct {
 	inMesh uint64
@@ -284,75 +265,6 @@ func getResultKeys(
 	return keys
 }
 
-// add filtering by resource type
-// note that metricToKey assumes the label ordering (namespace, name)
-func promGroupByLabelNames(resource *pb.Resource) model.LabelNames {
-	names := model.LabelNames{namespaceLabel}
-
-	if resource.Type != k8s.Namespace {
-		names = append(names, promResourceType(resource))
-	}
-	return names
-}
-
-// add filtering by resource type
-// note that metricToKey assumes the label ordering (namespace, name)
-func promDstGroupByLabelNames(resource *pb.Resource) model.LabelNames {
-	names := model.LabelNames{dstNamespaceLabel}
-
-	if isNonK8sResourceQuery(resource.GetType()) {
-		names = append(names, promResourceType(resource))
-	} else if resource.Type != k8s.Namespace {
-		names = append(names, "dst_"+promResourceType(resource))
-	}
-	return names
-}
-
-// query a named resource
-func promQueryLabels(resource *pb.Resource) model.LabelSet {
-	set := model.LabelSet{}
-	if resource.Name != "" {
-		set[promResourceType(resource)] = model.LabelValue(resource.Name)
-	}
-	if shouldAddNamespaceLabel(resource) {
-		set[namespaceLabel] = model.LabelValue(resource.Namespace)
-	}
-	return set
-}
-
-// query a named resource
-func promDstQueryLabels(resource *pb.Resource) model.LabelSet {
-	set := model.LabelSet{}
-	if resource.Name != "" {
-		if isNonK8sResourceQuery(resource.GetType()) {
-			set[promResourceType(resource)] = model.LabelValue(resource.Name)
-		} else {
-			set["dst_"+promResourceType(resource)] = model.LabelValue(resource.Name)
-			if shouldAddNamespaceLabel(resource) {
-				set[dstNamespaceLabel] = model.LabelValue(resource.Namespace)
-			}
-		}
-	}
-
-	return set
-}
-
-// determine if we should add "namespace=<namespace>" to a named query
-func shouldAddNamespaceLabel(resource *pb.Resource) bool {
-	return resource.Type != k8s.Namespace && resource.Namespace != ""
-}
-
-// query for inbound or outbound requests
-func promDirectionLabels(direction string) model.LabelSet {
-	return model.LabelSet{
-		model.LabelName("direction"): model.LabelValue(direction),
-	}
-}
-
-func promResourceType(resource *pb.Resource) model.LabelName {
-	return model.LabelName(resource.Type)
-}
-
 func buildRequestLabels(req *pb.StatSummaryRequest) (labels model.LabelSet, labelNames model.LabelNames) {
 	// labelNames: the group by in the prometheus query
 	// labels: the labels for the resource we want to query for
@@ -468,14 +380,6 @@ func processPrometheusMetrics(req *pb.StatSummaryRequest, results []promResult, 
 	return basicStats
 }
 
-func extractSampleValue(sample *model.Sample) uint64 {
-	value := uint64(0)
-	if !math.IsNaN(float64(sample.Value)) {
-		value = uint64(math.Round(float64(sample.Value)))
-	}
-	return value
-}
-
 func metricToKey(req *pb.StatSummaryRequest, metric model.Metric, groupBy model.LabelNames) rKey {
 	// this key is used to match the metric stats we queried from prometheus
 	// with the k8s object stats we queried from k8s
@@ -556,32 +460,4 @@ func checkContainerErrors(containerStatuses []apiv1.ContainerStatus, containerNa
 		}
 	}
 	return errors
-}
-
-func isInvalidServiceRequest(selector *pb.ResourceSelection, fromResource *pb.Resource) bool {
-	if fromResource != nil {
-		return fromResource.Type == k8s.Service
-	} else {
-		return selector.Resource.Type == k8s.Service
-	}
-}
-
-func (s *grpcServer) queryProm(ctx context.Context, query string) (model.Vector, error) {
-	log.Debugf("Query request:\n\t%+v", query)
-
-	// single data point (aka summary) query
-	res, err := s.prometheusAPI.Query(ctx, query, time.Time{})
-	if err != nil {
-		log.Errorf("Query(%+v) failed with: %+v", query, err)
-		return nil, err
-	}
-	log.Debugf("Query response:\n\t%+v", res)
-
-	if res.Type() != model.ValVector {
-		err = fmt.Errorf("Unexpected query result type (expected Vector): %s", res.Type())
-		log.Error(err)
-		return nil, err
-	}
-
-	return res.(model.Vector), nil
 }
